backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
Pillow==9.0.0
gunicorn==20.0.4